+++ sshd-core/src/main/java/org/apache/sshd/server/global/TcpipForwardHandler.java
@@ -51,7 +51,9 @@ public class TcpipForwardHandler extends AbstractConnectionServiceRequestHandler
     }
 
     @Override
-    public Result process(ConnectionService connectionService, String request, boolean wantReply, Buffer buffer) throws Exception {
+    public Result process(
+            ConnectionService connectionService, String request, boolean wantReply, Buffer buffer)
+                throws Exception {
         if (!REQUEST.equals(request)) {
             return super.process(connectionService, request, wantReply, buffer);
         }