+++ CHANGES.md
@@ -89,3 +89,6 @@ client.setSignatureFactories(
 
 * [SSHD-897](https://issues.apache.org/jira/browse/SSHD-897) - The default CLI code automatically tries to detect the PTY settings to use
 if opening a shell or command channel.
+
+* [SSHD-901](https://issues.apache.org/jira/browse/SSHD-901) - Added capability to request a reply for the `keepalive@...` heartbeat request
+in order to avoid client-side session timeout due to no traffic from server.